+++ models/OrdemServicoServico.cs
@@ -8,7 +8,7 @@ public class OrdemServicoServico
 
     public string ServicoNome { get; set; }
 
-    public decimal Quantidade { get; set; }
+    public int Quantidade { get; set; } // old: decimal
 
     public decimal PrecoUnitario { get; set; }
